[SAP] Replace CapacityFilter
This patch reworks the CapacityFilter which uses the new
utils.calculate_capacity_factors to determine if the
virtual free space available.

The main issue is that the vmware driver does lazy creates.
This causes an issue of over reporting of free space by
the backend, because the space hasn't been consumed yet.
So the amount of free space is not accurate with respect to
how much has been allocated by cinder.  This updated calculate
capacity factors as well as the SAPCapacityFilter accounts for
the virtual free space by using the cinder allocated capacity
tracking.  If the free space is reported less than what cinder
thinks should be available, then the reported free space is used.

This relies on an acurate reporting of the allocated capacity
by the driver.  We know there is an issue with allocated capacity
not being reported correctly for migrated volumes, as well as
accounting for existing volumes at startup.  The startup issue
should be solved with this PR:
https://github.com/sapcc/cinder/pull/117

Will have to do a folllow up to account for updating the allocated
capacity for migrated volumes.