[SAP] Added sap_affinity_filter.py
This patch adds the SAPDifferentBackendFilter and SAPSameBackendFilter
classes.  These classes are the affinity/anti-affinity classes to check
and ensure that the same/different netapp backends are used during
volume provisioning requests.

The same cinder api is used for passing in the scheduler hint.
cinder volume create --name foo --hint same_host=<UUID>
cinder volume create --name foo --hint different_host=<UUID>

If a datastore has a custom attribute set on it with the name of
netapp_fqdn then that value is compared in the filters to ensure
the same/different datastore is chosen, depending on the scheduler
hint passed in.