[SAP] ensure affinty/antiaffinity is maintained at migration time
This patch adds the volume create time scheduler hint to the
volume metadata and uses that metadata at volume_migrate time.

This ensures that when a volume is migrated via normal cinder migrate
call, that the affinity/anti-affinity is still honored.

The volume metadata keys of
scheduler_hint_same_host
scheduler_hint_different_host

is not editable or creatable by the API.

[SAP] fix the metadata for affinity/anti-affinity

This fixes a problem of storing the volume uuids in the metadata
{'scheduler_hint_same_host':
   'f,0,5,1,d,d,d,3,-,b,2,b,4,-,4,5,2,2,-,b,0,8,7,-,6,4,c,d,8,f,a,5,c,7,a,3'}

Change-Id: I96dfff2a866dcb6e48e6f6a0b56419492936bdd2