[SAP] ensure affinty/antiaffinity is maintained at migration time
This patch adds the volume create time scheduler hint to the
volume metadata and uses that metadata at volume_migrate time.

This ensures that when a volume is migrated via normal cinder migrate
call, that the affinity/anti-affinity is still honored.

The volume metadata keys of
scheduler_hint_same_host
scheduler_hint_different_host

is not editable or creatable by the API.