[SAP] Update allocated_capacity on remote host
This patch calls a new RPCAPI to tell the new remote host
that it needs to update it's allocated_capacity_gb after a
volume migration is successful.

This ensures that migrating between 2 different cinder volume
services can update the allocated_capacity_gb after a migration
from one host to another.   This happens when a volume is moved
from one shard to another.