[SAP] Update allocated_capacity on remote host
This patch calls a new RPCAPI to tell the new remote host
that it needs to update it's allocated_capacity_gb after a
volume migration is successful.

This ensures that migrating between 2 different cinder volume
services can update the allocated_capacity_gb after a migration
from one host to another.   This happens when a volume is moved
from one shard to another.

[SAP] Fix for allocated_capacity_gb not being updated

This patch puts a fix in for the scheduler's copy of the
backend/pool's allocated_capacity_gb while doing a migrate
by connected.

This patch also reorders the RPC call to update the
allocated_capacity_gb inside the volume manager's migrate_volume call.

This will ensure:
1) The scheduler's copy of allocated_capacity_gb for a backend/pool is
updated for a migrate_by_connector in between the volume manager's
get_volume_stats() calls.

2) after the volume manager calls get_volume_stats on a driver the
allocated_capacity_gb has been updated from the last migration.

These issues were discovered by an overprovisioning of a few datastores
in eu-de-2, when a customer created 10+ volumes quickly and attached all
of them to a VM.  The volumes were all created on 1 shard, and then due
to the attach call, they were all migrated to the shard where the nova
vm lived, all within a short period of time.

[SAP] Fix rpc call to remote host to update host capacity

This patch fixes a problem with the volume
rcpapi.update_migrate_volume_capacity()
It was always pulling the remote host to call from the
volume.service_topic_queue, which is always the current host entry, not
the new host that the volume is migrating to.  This is how we were
getting overcommited against remote datastores during volume attachments
inducing a migration to another vcenter.  The remote wasn't actually
getting called to update it's allocated_capacity_gb.

[SAP] update the new remote host allocated_capacity_gb

This patch fixes a problem updating the allocated_capacity_gb
for the destination cinder host during volume migration.

We recently made a change to when we update the allocated_capacity_gb
during volume migration, which allowed an over allocation.  The
remove rpc call to update the destination host was moved to before
the call to migrate the volume to ensure that the scheduler had updated
stats for the destination, since the migration can take a long time.

The problem with this move is that the volume host entry wasn't updated
at the time of the call to update the remote destination host.  The
update was being called on the source host, so there was effectively
no change since the src and destination host was the same during both
calls to update_allocated_capacity_gb.  This patch fixes that.

Added a host entry to the rpc call to force the update on that
particular host value.

Change-Id: Ic1021bfd8763db5c969928614075c4af2b43532e