[SAP] Fix migration by connector filter_properties
This patch fixes the migration by connector with affinity/anti-affinity
filter_properties.  The previous code was doing an rpcapi call to the
volume manager to migrate the volume with the filter properties.
At this point the scheduler has already picked a destination host, so
the filter_properties being passed to the volume manager is useless.

This patch sends the filter_properties through the scheduler to help it
pick the right backend before calling the volume manager to migrate
the volume to the new backend.