[SAP] Backup: Limit number of concurent operations
This is a backport from upstream master:
30c2289c9b0456d3783f01e3d65985ed1b09976a
https://github.com/openstack/cinder/commit/30c2289c9b0456d3783f01e3d65985ed1b09976a

This patch adds the backup_max_operations configuration option that
limits the number of concurrent backup and restore operations.

These operations can consume large amounts of RAM, so we should limit
them according to our system resources to avoid processes being killed
because we run out of memory.

The default is 15 concurrent operations, which could use up to 85GB
of RAM(in case of ceph which uses the maximum memory for this operation).
More details are added in the document.

Closes-Bug: #1865011
Co-authored-by: Rajat Dhasmana <rajatdhasmana@gmail.com>