[SAP] Added pool_state tracking. Updated backend_state
This patch adds the ability to track if a datastore is available
to be used in get_volume_stats.  If a datastore is in maintenance
mode, then it can't accept volume provisioning requests.  So the
driver now marks the datastore/pool as 'down'.  When the datastore
comes back out of maintenance mode, then the pool_state is changed to
'up'.

Currently, Cinder doesn't have a built in mechanism for dealing with a
pool_state being down, but we can simulate it by adding an extra spec
setting in all volume types of pool_state='up'.  That way the
capability filter will filter out all pools that don't have that
capability coming from the driver.

This patch also changes where the backend_state is set so that
the scheduler can mark the service as down correctly.