[SAP] update the new remote host allocated_capacity_gb
This patch fixes a problem updating the allocated_capacity_gb
for the destination cinder host during volume migration.

We recently made a change to when we update the allocated_capacity_gb
during volume migration, which allowed an over allocation.  The
remove rpc call to update the destination host was moved to before
the call to migrate the volume to ensure that the scheduler had updated
stats for the destination, since the migration can take a long time.

The problem with this move is that the volume host entry wasn't updated
at the time of the call to update the remote destination host.  The
update was being called on the source host, so there was effectively
no change since the src and destination host was the same during both
calls to update_allocated_capacity_gb.  This patch fixes that.

Added a host entry to the rpc call to force the update on that
particular host value.