[SAP] Fix the shard filter for group creation
Temptest tests found a bug in the shard filter where creating
a cinder group from an existing group or group snapshot always failed
with not finding any hosts.  This is due to the shard filter not getting
much of any information about the request.

This patch adds the project_id of the group being created to ensure that
the shard filter can work properly.