Allow snapshots to be independent
This patch allows a snapshot creation request go through the
scheduler to pick a pool for the snapshot to live on.
The backend picked by the scheduler is added to the snapshot
metadata.  This metadata field '__cinder_internal_backend'
is filtered out of requests fetching the snapshot information,
so end users will never see it.

To enable this ability a new config option is added to the
scheduler and is defaulted to be False or disabled.

sap_allow_independent_snapshots

This allows creating a snapshot from a volume to happen on a
completely different pool than the source volume.

This patch also alters create volume from snapshot, to allow a volume
to be created from a different pool than the source volume that
the snapshot was created from.

All of this is not how upstream works or allows.

Fix snapshot view filtering

This patch fixes an issue iterating over the metadata
from the snapshot views API.  The code is supposed to filter out
and remove the __cinder_internal keys in the metadata.  This patch
looks for the keys to remove first and then deletes them instead
of iterating over the metadata array and trying to delete them while
iterating over the metadata, which causes a python exception.

[SAP] Use snapshot hidden backend for new volumes

When creating a volume from a snapshot, the hidden
__cinder_internal_backend metadata on the snapshot will
be used for the host for the volume creation.  We do this now
instead of falling back to the snapshot's source volume host can
be different than it was when the snapshot was changed, because
the source volume can be migrated now.

This ensures that the new volume from snapshot can copy the bits
from the snapshot (clone).

Change-Id: Iece6ac694c22a83fe08c3fa14114fafd68218429