[SAP] Allow snapshots to be independent
This patch allows a snapshot creation request go through the
scheduler to pick a pool for the snapshot to live on.
The backend picked by the scheduler is added to the snapshot
metadata.  This metadata field '__cinder_internal_backend'
is filtered out of requests fetching the snapshot information,
so end users will never see it.

To enable this ability a new config option is added to the
scheduler and is defaulted to be False or disabled.

sap_allow_independent_snapshots

This allows creating a snapshot from a volume to happen on a
completely different pool than the source volume.

This patch also alters create volume from snapshot, to allow a volume
to be created from a different pool than the source volume that
the snapshot was created from.

All of this is not how upstream works or allows.