[SAP] Fix snapshot view filtering
This patch fixes an issue iterating over the metadata
from the snapshot views API.  The code is supposed to filter out
and remove the __cinder_internal keys in the metadata.  This patch
looks for the keys to remove first and then deletes them instead
of iterating over the metadata array and trying to delete them while
iterating over the metadata, which causes a python exception.