Mark pool as down if datastore draining
This patch looks for a custom attribute 'cinder_state' on a datastore
at get_volume_stats() time.  If the custom attribute value is 'drain',
then the datastore is marked as 'draining' and cinder shouldn't use it
for provisioning new requests.  The datastore will be marked as down
in the pool stats for that datastore.