[SAP] Mark pool as down if datastore draining
This patch looks for a custom attribute 'cinder_state' on a datastore
at get_volume_stats() time.  If the custom attribute value is 'drain',
then the datastore is marked as 'draining' and cinder shouldn't use it
for provisioning new requests.  The datastore will be marked as down
in the pool stats for that datastore.

[SAP] Fix vmdk stats reporting

There is a failure for looking at the cinder_pool_state for a datastore
when the state is None (not set).  This patch ensures that the custom
attribute is set to something before trying to call .lower() on the
string.

Change-Id: Ia57d7f74b6233632327f754eacb02be10c2daeb8