[SAP] Update shard filter to filter only on vmware
This patch adds a simple check against the pool in the shard filter
to only filter if the pool is reporting as a vendor_name of 'VMware',
which is true only for vmdk.py and fcd.py drivers.

If we deploy a netapp cinder driver, this will pass the shard filter
as there is no reason to shard.