[SAP] ensure snapshot clones accounted for
This patch adds the ability in the volume manger to check if a backend
has clones as snapshots and account for the space consumed on the pool.

All of our snapshots are clones and the volume manager doesn't account
for the space consumed on the pool/datastore.  Only the scheduler
accounts for that space consumed on the pool currently. With this patch
the volume manager will check all snapshots in the DB and properly
account for the space used on the pool. This is to help prevent
overcommit on a pool.