[SAP] Use snapshot hidden backend for new volumes
When creating a volume from a snapshot, the hidden
__cinder_internal_backend metadata on the snapshot will
be used for the host for the volume creation.  We do this now
instead of falling back to the snapshot's source volume host can
be different than it was when the snapshot was changed, because
the source volume can be migrated now.

This ensures that the new volume from snapshot can copy the bits
from the snapshot (clone).