[SAP] FCD Driver
Patches for making the FCD driver work.

Change FCD driver to use volume.name attribute for directory naming, also use it for temp directory.
This change will help ops to find fcd objects on the datastore, as there is no longer any searchable object in vcenter....

[SAP] Fix vmware fcd driver delete_volume on failure

This patch catches the VIMException when an fcd volume isn't found
during a delete_volume.  This is considered a success.

[SAP] Update FCD driver

This patch updates the FCD driver to work in our deployment.

1. Create volumes on the requested datastore from the scheduler.

2. This patch refactors the driver name used in the location information
for a volume metadata.  This also ensures that the migrate_volume
only works on the appropriate driver for the driver_name.
3. FCD changes that enable cross vcenter migration for FCD based volumes.
4. FCD copy image to volume working
5. Add cross vc unattached migration for pre 8.0u3 vc versions
6. Cleanup after fcd delete, we dispose of the empty folder
7. Fixed retype to fcd.
8. Add new SAPFCD filter

This new filter is to help ensure that a cross vcenter migration
lands on the same pool as the originating pool only for a
volume migration operation for the FCD driver.  This will result
in no data being moved as the datastores are actually the same
netapp backed nfs volume.

9. Fixed a call to _get_temp_image_folder

The fcd driver has to find the location of the folder
based on the volume, which contains the host entry, so
this patch renames _get_temp_image_folder in the fcd driver
to _get_temp_image_folder_from_volume to distinguish itself
from the vmdk driver.

10. store provider_location as ds name

This patch updates the FCD driver to store the
provider_location as the datastore name instead
of the datastore moref.  This is to solve a
potential future issue where an nfs datastore is remounted.
When an nfs datastore is remounted the moref id changes, even
with the same datastore name.

This patch also hard codes the fcd driver to create snapshot
as a clone of the original volume.

11. FCD clone while attached

This patch addresses an issue with vcenter failing to clone
an FCD based volume while that volume is attached to a vm.

12. FCD backup volumes working

This patch refactors the initialize_connection a bit between
the parent class vmdk.py and fcd.py to get volume backup
to work

13. [SAP] Got backup restore working for FCD

This patch changes the terminate_connection call in
the volume manager to fetch the volume object instead
of the volume record from the DB.  The FCD driver needs
the volume object to update the provider location
for the restored volume.

This patch also updates the terminate_connection for the
fcd driver to properly move and rename the restored write handle
vmdk to the correct volume folder and vmdk name.

14. fixed unit tests

This patch fixes the unit tests needed for fcd.

Don't create shadow vm on initialize_connection not from backup pod

Fix unit test

Update fcd.py fix initiate_connection

Shadow vm had a duplicate vmdk file, we can pass a parameter to create a diskless backing to create_backing call

Update fcd.py terminate_connection

Delete the shadow vm on read kind of terminate_connection

[SAP] Fix retype to fcd

This patch fixes a minor issue when retyping an vmdk volume to
an fcd volume.  the provider_location for an FCD volume now needs
to the the datastore name in the path instead of the vmware datastore
id.

Change-Id: I8259ebcb42e3a5a0874e00360413974e393c5779

[SAP] update independent snapshot code

This patch adds the additional check for the volume type
containing the key/value of
independent_snapshot = True as well as the
CONF.sap_allow_independent_snapshot.

Previously, the only driver enabled anywhere was the vmdk.py
driver which needs independent snapshots created as they are
full clones.  FCD driver doesn't need/have independent snaps.

This prevents the FCD volume snapshots from getting placed on
a vmware (vmdk.py) datastore.

[SAP] Fix snapshot accounting

This patch makes a fix for snapshot accounting on a pool
when the backend doesn't support independent snapshots.

Change-Id: I9be101e45a1413343da375fe7030a2d23fee4282

[SAP] Enable independent snapshots for fcd

This patch looks for the global independent snapshots
setting and if that's on, enable it for the fcd driver.
independent snaps are for backends that do clones instead
of snapshots for their snapshots.  This allows the volume
manager to track the storage usage on the pool where the
snapshot exists.

Change-Id: I07332a0df2b49ad97a9c0cc8fc074144b66171df

[SAP] Add online-retype to fcd (#204)

* Add online-retype to fcd
* Change api_version call in update_server_volume to support bdm update without swap
* Rework retype code to support vsphere 7

Update attachments after retype of in-use volume

Remove config from connection_info as it leaks cinder credentials to nova, fix dc_obj value

Fix cinder connection_info to be a flat structure, as nova already has a _translate_attachment_ref function in nova/volume/cinder.py to transform the attachments

[SAP] Fix retyping lazy vmdk to fcd volume

This patch fixes an issue with retyping from a vmdk volume to
an fcd volume.  If the vmdk volume is lazy created and hasn't
been attached yet, there was no backing, so we have to first
create the backing so we can migrate it to an fcd volume.

Change-Id: Ibfd81e9bcdaacaf20adaa73b4d69f0b170bf61b8

[SAP] fix backup restore for FCD

This patch is fixes a problem with restore from backup
for the FCD driver.

[SAP] fix backup restore for FCD part2

This patch is fixing the different behaviour of vsphere 8 around file locking,
fixes a problem with restore from backup
for the FCD driver.

We had to add a sleep call to give vcenter time to release the
lock on the vmdk file.

Change-Id: Ib077cd5a54f4f4288d9a5d965c3e0d4a0bdca104

[SAP] fixes for migration of an fcd volume.

This patch contains fixes for migrating volumes between different
fcd based datastores.

update_fcd_policy requires the datastore in (_type:"Datastore,value:"mo-id") format

Fix previous commit

Fix unit tests and fix the migration of unattached volumes.

Change-Id: Ic82d7fc4593cb33d742d588bc259515748273baa

[SAP] Allow migration for vmdk -> fcd

This patch updates the vmdk.py driver to allow migration of a
vmdk based volume to fcd.

Change-Id: I1b79d54398cda6aa8e47bbd0c001da214320d487