SAP: Add affinity UUID validation
This patch adds some basic UUID volume validation for the cinder
create volume API request when passing in scheduler hints for
volume affinity or anti-affinity.   The API now ensures that the
UUIDs are valid cinder volumes.  The UUID must be a valid cinder
volume UUID for the create call to work.