SAP: ShardFilter don't k8s filter on migration
This patch adds a check in the ShardFilter to not do any k8s
filtering on a volume migration.  For volumes in place prior to
the k8s filtering feature of the ShardFilter, the shard that a volume
currently lives on can get filtered out due to the k8s filtering
and the sql query odering.  This could prevent a volume from being moved
to another datastore in the same shard.

This patch adds an 'action' to the volume migration request spec
with the value of 'volume_migrate' and the ShardFilter looks for it
to tell it to not filter k8s volume hosts out.