ShardFilter: allow existing K8S clusters with more than 1 shard
Extend the SQL query to return all matching hosts, ordered desc.
For new volume creations we allow it in the dominant shard only,
however for existing volumes we allow the operation in all the
shards of the K8S cluster. This is a backward-compatibility with
existing K8S cluster spanning multiple shards in the same AZ.

Additionally, this commit fixes the availability_zone value to be
taken from filter_properties, which is always seeded by
populate_filter_properties()

For better debugging, adding a debug log when a backend is
filtered out by the K8S logic.

Change-Id: I42d526e8b1335e9025543a424f9203858a44663e