[Pure Storage] Driver API version upgrade
With the upcoming deprecation of REST 1.x used by this driver to
comminucate with the FlashArray, this patch is to convert the
driver to use REST 2.x calls, which means changing the required
Python SDK from ``purestorge`` (REST 1.x) to ``py-pure-client``
(REST 2.x). The minimum version of ``py-pure-client`` is
defined to be 1.47.0.

This change does not affect any of the functionality of the driver
but does affect the minimum required FlashArray operating system,
increasing this to Purity//FA 6.1.0.

This change will also allow for the addition of new features
that were not supported by the REST 1.x APIs.

Documentation changes are made to reflect the change in required
SDK.

All unit tests are updated to reflect the changes in the driver
and the use of new SDK calls.

Change-Id: I3deae1cc6bbe5ef5820224f72bb02565677e1ade