[SAP] change name for temp backing
This patch updates the name of a temporary backing that is created
during the cloning an attached volume.   This prefixes the name
with an obvious string (TEMP_BACKING-<uuid>), so that we can see
orphaned temporary backings easily on vcenter.

This patch also updates the extra_config to remove a few items
in the temp backing.