[SAP] Account for snapshot capacity during create
This patch updates the local accounting for the allocated capacity
for a snapshot that has a SAP_HIDDEN_BACKEND_KEY in the metadata.
That key is only there if independent snapshots are allowed.  It's
assumed that the snapshot is a full clone as well, when the
backend allows for independent snaps.

Change-Id: I3cbf5a2e59f59fa1594b494d0e4b5c16bcbde62a