[SAP] refactor allocated capacity tracking
This patch reworks some of the code for allocated capacity tracking.
There were a few places that was touching the pool's
allocated_capacity_gb.  This patch refactors that into a single
function in the volume manager.

This patch also adds a check for updating the pool's capacity
tracking for snapshots and ensures that the driver that owns
the snapshot is creating clones.

This patch also adds code to decrement the capacity when
a cloned snapshot is deleted.

Change-Id: Ia32d7582ea21bc32d5eff6ea73d618cf7e65259a