[SAP] Fix retyping lazy vmdk to fcd volume
This patch fixes an issue with retyping from a vmdk volume to
an fcd volume.  If the vmdk volume is lazy created and hasn't
been attached yet, there was no backing, so we have to first
create the backing so we can migrate it to an fcd volume.

Change-Id: Ibfd81e9bcdaacaf20adaa73b4d69f0b170bf61b8
(cherry picked from commit 2258d313106b7f96d34522871eaf3c4f87b56801)