[SAP] Enable independent snapshots for fcd
This patch looks for the global independent snapshots
setting and if that's on, enable it for the fcd driver.
independent snaps are for backends that do clones instead
of snapshots for their snapshots.  This allows the volume
manager to track the storage usage on the pool where the
snapshot exists.

Change-Id: I07332a0df2b49ad97a9c0cc8fc074144b66171df