[SAP] fix backup restore for FCD
This patch is fixes a problem with restore from backup
for the FCD driver.

[SAP] fix backup restore for FCD part2

This patch is fixing the different behaviour of vsphere 8 around file locking,
fixes a problem with restore from backup
for the FCD driver.

We had to add a sleep call to give vcenter time to release the
lock on the vmdk file.

Change-Id: Ib077cd5a54f4f4288d9a5d965c3e0d4a0bdca104
(cherry picked from commit bece57f89bde354dccfb9c6f65e3f76ad7b85f40)