[SAP] Start work on FCD QOS setting
This patch is the start of the work needed to enable setting
qos settings for standard_hdd volume type volumes for netapp .vmdk
files.

This patch simply lays out the logic for finding extra spec setting
that tells the driver that the volume to be created is for standard hdd.
Added todos in comments to actually do the work.

Change-Id: Ie0f59716c9f28076d65dc3c0580c836f6879f90d